# internal/apidoc/swagger.yaml
basePath: /
definitions:
  api.estimateRequest:
    properties:
      datasource:
        description: id or name
        type: string
      session_id:
        description: optional; links queries from one AI conversation
        type: string
      sql:
        description: SQL statement to estimate
        type: string
    type: object
  api.loginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  api.loginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/api.meResponse'
    type: object
  api.meResponse:
    properties:
      attributes:
        additionalProperties:
          type: string
        type: object
      display_name:
        type: string
      id:
        type: string
      roles:
        items:
          type: string
        type: array
      username:
        type: string
    type: object
...
  api.nl2sqlRequest:
    properties:
      datasource:
        description: id or name
        type: string
      question:
        description: natural-language question
        type: string
      session_id:
        description: optional; links queries from one AI conversation
        type: string
      sql_hint:
        description: optional hand-written SQL to prefer over generation
        type: string
    type: object
  api.queryRequest:
    type: object
host: localhost:8080
info:
  contact:
    name: Aegis
    url: https://github.com/wisonwang/Aegis
  description: Aegis — AI Data Supply Gateway. 受治理的 DataAPI + 面向 AI Agent 的 MCP 服务。默认拒绝：每条查询都经过权限引擎重写与脱敏。
  license:
    name: MIT
  termsOfService: https://github.com/wisonwang/Aegis
  title: Aegis DataAPI
  version: "1.0"
paths:
...
  /admin/api/datasources:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: admin List Data Sources
      tags:
      - admin
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: admin Create Data Source
      tags:
      - admin
...
  /api/v1/datasources:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List registered datasources
      tags:
      - dataapi
  /api/v1/datasources/{id}/catalog:
    get:
      parameters:
      - description: datasource id or name
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Governed, semantically enriched schema catalog
      tags:
      - dataapi
...
  /api/v1/datasources/{id}/nl2sql:
    post:
      consumes:
      - application/json
      parameters:
      - description: datasource id or name
        in: path
        name: id
        required: true
        type: string
      - description: nl2sql request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.nl2sqlRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Natural-language question to governed SQL
      tags:
      - dataapi
  /api/v1/datasources/{id}/query/estimate:
    post:
      consumes:
      - application/json
      parameters:
      - description: datasource id or name
        in: path
        name: id
        required: true
        type: string
      - description: estimate request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.estimateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Estimate query cost/risk without executing
      tags:
      - dataapi
  /api/v1/datasources/{id}/tables:
    get:
      parameters:
      - description: datasource id or name
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List accessible tables on a datasource
      tags:
      - dataapi
  /api/v1/datasources/{id}/tables/{table}:
    get:
      parameters:
      - description: datasource id or name
        in: path
        name: id
        required: true
        type: string
      - description: table name
        in: path
        name: table
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Describe table columns (governed)
      tags:
      - dataapi
  /api/v1/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.loginResponse'
      summary: Authenticate and obtain a JWT
      tags:
      - auth
  /api/v1/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.meResponse'
      security:
      - BearerAuth: []
      summary: Get the current principal's profile
      tags:
      - auth
...
  /api/v1/query:
    post:
      consumes:
      - application/json
      parameters:
      - description: query request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.queryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Execute a governed query
      tags:
      - dataapi
...
securityDefinitions:
  BearerAuth:
    description: '"Bearer <JWT> — 通过 POST /api/v1/login 获取"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"